aggregating all data generation projects into one
