--- environment.yaml
@@ -1,13 +1,0 @@
-name: bbhnet-base
-channels:
-  - pytorch
-  - conda-forge
-dependencies:
-  - python=3.9
-  - gwpy=2.1.0
-  - astropy<5.0.0
-  - python-ldas-tools-framecpp
-  - python-nds2-client
-  - pytorch
-  - torchvision
-  - cudatoolkit=11.3
