--- libs/injection/bbhnet/injection/__init__.py
@@ -1,2 +1,0 @@
-from .injection import generate_gw, inject_waveforms
-from .priors import end_o3_ratesandpops, nonspin_bbh
