--- libs/injection/pyproject.toml
@@ -1,25 +1,0 @@
-[tool.poetry]
-name = "bbhnet.injection"
-version = "0.0.1"
-description = "Utilities for generating BBH signals using bilby"
-authors = ["Will Benoit <william.benoit@ligo.org>",]
-license = "MIT"
-packages = [
-    {include = "bbhnet"}
-]
-
-[tool.poetry.dependencies]
-python = ">=3.8,<3.11"
-bilby = "<1.2"
-lalsuite = "^7.4"
-
-"bbhnet.base" = {path = "../base", extras = ["numpy", "gwpy"], develop = true}
-
-
-[tool.poetry.dev-dependencies]
-pytest = "^6.2"
-ipykernel = "^6.15.1"
-
-[build-system]
-requires = ["poetry>=1.2"]
-build-backend = "poetry.masonry.api"
