+++ projects/sandbox/datagen/pyproject.toml
@@ -1,0 +1,31 @@
+[tool.poetry]
+name = "datagen"
+version = "0.0.1"
+description = "Dataset generation utilities"
+authors = [
+    "Ethan Marx <emarx@mit.edu>",
+    "Will Benoit <william.benoit@ligo.org>",
+    "Alec Gunny <alecg@mit.edu>"
+]
+license = "MIT"
+
+[tool.poetry.scripts]
+generate-background = "datagen.scripts.background:main"
+generate-glitches = "datagen.scripts.glitches:main"
+generate-waveforms = "datagen.scripts.waveforms:main"
+generate-timeslides = "datagen.scripts.timeslides:main"
+
+[tool.poetry.dependencies]
+python = ">=3.8,<3.11"
+typeo = {git = "https://github.com/ML4GW/typeo.git", branch = "main"}
+ml4gw = {path = "../../../ml4gw", develop = true}
+bilby = "<1.2"  # set this so ml4gw doesn't update it on accident
+
+"bbhnet.logging" = {path = "../../../libs/logging", develop = true}
+
+[tool.poetry.group.dev.dependencies]
+pytest = "^6.2"
+
+[build-system]
+requires = ["poetry>=1.2"]
+build-backend = "poetry.masonry.api"
