--- projects/sandbox/generate_background/pyproject.toml
@@ -1,23 +1,0 @@
-[tool.poetry]
-name = "generate_background"
-version = "0.1.0"
-description = "MIT"
-authors = ["ethan.marx <ethan.marx@ligo.org>"]
-license = "MIT"
-
-[tool.poetry.dependencies]
-python = ">=3.8,<3.11"
-typeo = {git = "https://github.com/ML4GW/typeo.git"}
-
-"bbhnet.logging" = {path = "../../../libs/logging", develop = true}
-"bbhnet.base" = {path = "../../../libs/base", extras = ["gwpy", "numpy"], develop = true}
-
-[tool.poetry.group.dev.dependencies]
-pytest = "^7.1.3"
-
-[tool.poetry.scripts]
-generate-background = "generate_background:main"
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
