--- projects/sandbox/generate_glitches/environment.yaml
@@ -1,11 +1,0 @@
-name: bbhnet-generate_glitches
-channels:
-  - conda-forge
-dependencies:
-  - python=3.9
-  - gwpy
-  - python-ldas-tools-framecpp
-  - python-nds2-client
-  - python-ligo-lw<1.8.0
-  - pyomicron
-  - omicron=2.4.2
