--- projects/sandbox/generate_glitches/pyproject.toml
@@ -1,28 +1,0 @@
-[tool.poetry]
-name = "generate_glitches"
-version = "0.0.1"
-description = "Generate a glitch dataset for training/validation"
-authors = ["Ethan Marx <emarx@mit.edu>",]
-license = "MIT"
-
-[tool.poetry.dependencies]
-python = ">=3.8,<3.11"
-dqsegdb2="<1.1.0"
-pre-commit = "^2.19.0"
-typeo = {git = "https://github.com/ML4GW/typeo.git"}
-
-"bbhnet.logging" = {path = "../../../libs/logging", develop = true}
-"bbhnet.base" = {path = "../../../libs/base", extras = ["numpy"], develop = true}
-
-# needed by gwpy in some pyomicron utility but not installed for some reason
-uproot = "^4.3.7"
-
-[tool.poetry.dev-dependencies]
-pytest = "^6.2"
-
-[tool.poetry.scripts]
-generate-glitches = "generate_glitches:main"
-
-[build-system]
-requires = ["poetry>=1.2"]
-build-backend = "poetry.masonry.api"
