--- projects/sandbox/generate_waveforms/pyproject.toml
@@ -1,33 +1,0 @@
-[tool.poetry]
-name = "generate_waveforms"
-version = "0.0.1"
-description = "Simulate BBH a waveform dataset for training/validation"
-authors = ["Will Benoit <william.benoit@ligo.org>",]
-license = "MIT"
-
-[tool.poetry.dependencies]
-python = ">=3.8,<3.11"
-typeo = {git = "https://github.com/ML4GW/typeo.git"}
-
-"bbhnet.injection" = {path="../../../libs/injection", develop = true}
-"bbhnet.logging" = {path="../../../libs/logging", develop = true}
-
-[tool.poetry.dev-dependencies]
-pytest = "^6.2"
-
-[tool.typeo]
-outdir = "${OUT_DIR}"
-n_samples = 10000
-prior_file = "./priors/nonspin_BBH.prior"
-
-# don't think we need the 
-# 4096 default specified
-sample_rate = 2048 
-waveform_duration = 8
-
-[tool.poetry.scripts]
-generate-waveforms = "generate_waveforms:main"
-
-[build-system]
-requires = ["poetry>=1.2"]
-build-backend = "poetry.masonry.api"
