--- projects/sandbox/timeslide_injections/O3b_GW_times.txt
@@ -1,35 +1,0 @@
-1256779567
-1256999739
-1257296855
-1257664691
-1258804397
-1258866165
-1259070047
-1259492747
-1259514944
-1260484270
-1260567236
-1260808298
-1261020955
-1261764316
-1262879936
-1263097407
-1264213229
-1264316116
-1264693411
-1265202095
-1265235995
-1265273710
-1265361792
-1265926102
-1266140673
-1266214786
-1266238148
-1266618172
-1266645879
-1267149509
-1267522652
-1267724187
-1267963151
-1268431094
-1268903511
