--- projects/sandbox/timeslide_injections/poetry.toml
@@ -1,2 +1,0 @@
-[virtualenvs]
-create = false
