--- projects/sandbox/timeslide_injections/pyproject.toml
@@ -1,30 +1,0 @@
-[tool.poetry]
-name = "timeslide_injections"
-version = "0.0.1"
-description = "Create timeslides with and without GWs injected"
-authors = ["Will Benoit <william.benoit@ligo.org>",]
-license = "MIT"
-
-[tool.poetry.scripts]
-timeslide_injections = "timeslide_injections.main:main"
-
-[tool.poetry.dependencies]
-python = ">=3.9,<3.11"
-bilby = "<1.2"
-typeo = {git = "https://github.com/ML4GW/typeo.git", branch = "main"}
-
-"bbhnet.injection" = {path="../../../libs/injection", develop=true}
-"bbhnet.io" = {path = "../../../libs/io/", develop = true}
-"bbhnet.logging" = {path = "../../../libs/logging/", develop = true}
-"bbhnet.parallelize" = {path = "../../../libs/parallelize/", develop = true}
-"ml4gw" = {path = "../../../ml4gw/", develop = true}
-
-
-[tool.poetry.dev-dependencies]
-pytest = "^6.2"
-line-profiler = "^3.5.1"
-ipykernel = "^6.15.0"
-
-[build-system]
-requires = ["poetry>=1.2"]
-build-backend = "poetry.masonry.api"
