+++ pyproject.toml
@@ -1,19 +1,3 @@
-[tool.poetry]
-name = "bbhnet"
-version = "0.0.1"
-description = "End-to-end BBHNet training, inference, and analysis repo"
-authors = [
-    "Alec Gunny <alec.gunny@ligo.org>",
-    "Will Benoit <william.benoit@ligo.org>",
-    "Ethan Marx <emarx@mit.edu>",
-]
-
-[tool.poetry.dependencies]
-python = ">=3.8,<3.10"
-
-[tool.poetry.group.dev.dependencies]
-pre-commit = "^2.16"
-
 [tool.black]
 line-length = 79
 include = '\.pyi?$'
@@ -31,7 +15,3 @@ exclude = '''
 known_first_party = ["bbhnet"]
 multi_line_output = 3
 include_trailing_comma = true
-
-[build-system]
-requires = ["poetry>=1.2"]
-build-backend = "poetry.masonry.api"
